fix: Update the iconfont component
