+++ ui/src/components/IconFont/index.tsx
@@ -1,6 +1,6 @@
 import { createFromIconfontCN } from "@ant-design/icons";
 const IconFont = createFromIconfontCN({
-  scriptUrl: "//at.alicdn.com/t/font_3023111_0s219ot3yzbb.js",
+  scriptUrl: "//at.alicdn.com/t/font_3023111_34dw75tq2bm.js",
 });
 
 export default IconFont;
