+++ ui/src/pages/Alarm/hooks/useAlarmEnums.ts
@@ -5,6 +5,23 @@ const useAlarmEnums = () => {
   const ChannelTypes = [
     { name: i18n.formatMessage({ id: "dingTalk" }), value: 1 },
   ];
-  return { ChannelTypes };
+  const AlarmStatus = [
+    {
+      status: 1,
+      label: "未开启",
+      color: "#108ee9",
+    },
+    {
+      status: 2,
+      label: "已开启",
+      color: "#87d068",
+    },
+    {
+      status: 3,
+      label: "正在报警",
+      color: "#f50",
+    },
+  ];
+  return { ChannelTypes, AlarmStatus };
 };
 export default useAlarmEnums;
